feat: moved ui components in separate folder
